% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_own_ETF.R
\name{calculate_etf}
\alias{calculate_etf}
\title{Calculate price of self-defined ETF of selected stocks}
\usage{
calculate_etf(etf_stocks, etf_name = "ETF")
}
\arguments{
\item{etf_stocks}{vector of stock symbols that are components of the ETF}

\item{eft_name}{name of the ETF}
}
\description{
Calculate price of self-defined ETF of selected stocks
}
