% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_close_price_to_EMA20.R
\name{compare_ema20}
\alias{compare_ema20}
\title{Display the percentage difference of stock close price to EMA20}
\usage{
compare_ema20(stock, n_days = 1000)
}
\arguments{
\item{stock}{stock symbol}

\item{n_days}{number of days to display}
}
\description{
Display the percentage difference of stock close price to EMA20
}
