% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{display_stocks}
\alias{display_stocks}
\title{plot chart of stocks}
\usage{
display_stocks(stocks, period = 350, description = NULL)
}
\arguments{
\item{stocks}{vector of stock symbols such as c("C", "AAPL")}

\item{period}{number of trading day to display}

\item{description}{description of the chart, stock symbol by default}
}
\description{
plot chart of stocks
}
