% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_stock_selection.R
\name{download_stock_data}
\alias{download_stock_data}
\title{Download historical stock prices}
\usage{
download_stock_data(stocks, past_days = 7)
}
\arguments{
\item{stocks}{a vector of stocks to select from}

\item{past_days}{integer, download data from a few days ago}
}
\value{
a data.table
}
\description{
Download historical stock prices
}
