% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_stock_selection.R
\name{get_active_stock}
\alias{get_active_stock}
\title{Get actively traded stocks}
\usage{
get_active_stock(stock_data)
}
\arguments{
\item{stock_data}{stock data downloaded with function download_stock_data()}
}
\value{
a vector of stocks
}
\description{
Get actively traded stocks whose average daily trading volume is above 100k
in the past week
}
