% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stock_selection.R
\name{get_all_stocks}
\alias{get_all_stocks}
\title{Get and clean all stocks listed in AMEX, NASDAQ, and NYSE}
\usage{
get_all_stocks()
}
\value{
a data.table
}
\description{
Get and clean all stocks listed in AMEX, NASDAQ, and NYSE
}
