% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stock_selection.R
\name{get_change}
\alias{get_change}
\title{Get price changes of stocks in last 5 trading days}
\usage{
get_change(stocks, keep_days = 3, past_days = 7)
}
\arguments{
\item{stocks}{a vector of stocks to select from}

\item{keep_days}{integer, keep data for the past few trading days}

\item{past_days}{integer, download data from a few days ago}
}
\value{
a data.table
}
\description{
To search for stocks whose price rose or fell beyond a certain
value in the past n trading days.
}
