% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_data.R
\name{get_financials}
\alias{get_financials}
\title{Download annual financial data}
\usage{
get_financials(stocks)
}
\arguments{
\item{stocks}{vector of stock symbols such as c("C", "AAPL")}
}
\value{
a data.table
}
\description{
Including revenue, margin, income, dividends, earnings,
cash flow, capital, ... ...
}
\examples{
aaa <- get_financials(c("AAPL", "C"))

}
