% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_stock_selection.R
\name{get_growth}
\alias{get_growth}
\title{Find growth stocks}
\usage{
get_growth(stocks, cut_off = 5)
}
\arguments{
\item{stocks}{: vector of stocks to search within}

\item{cut_off}{: cut off ratio of revenue growth percent}
}
\value{
a vector of stock symbol
}
\description{
search for stocks that have revenue growth rate above cut_off
each year in last three years
}
