% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_stock_selection.R
\name{get_jumper}
\alias{get_jumper}
\title{Find jumping stocks}
\usage{
get_jumper(stock_data, n_days = 1, win_cut = 10, lose_cut = -10)
}
\arguments{
\item{stock_data}{stock data downloaded with function download_stock_data()}

\item{n_days}{past n trading days}

\item{win_cut, lose_cut}{cutoff percent to be considered as jumping stocks}
}
\value{
a list of winners and losers
}
\description{
To search for stocks whose price rose or fell beyond a certain
value in the past n trading days.
}
