% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_data.R
\name{get_revenues}
\alias{get_revenues}
\title{Get annual revenues and caculate rate of changes}
\usage{
get_revenues(stocks)
}
\arguments{
\item{stocks}{vector of stock symbols such as c("C", "AAPL")}
}
\value{
a data.table
}
\description{
Get annual revenues and caculate rate of changes
}
\examples{
aaa <- get_revenues(c("AAPL", "C"))

}
