% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_stock_selection.R
\name{get_reversals}
\alias{get_reversals}
\title{Find reversal stocks}
\usage{
get_reversals(stocks, n1 = 30, n2 = 180, fraction = 0.3)
}
\arguments{
\item{stocks}{a vector of stocks to select from}

\item{n1, n2}{n1 and n2 days back from today}

\item{fraction}{the ratio of the trading range between n1 and n2 days to
that of the past two years}
}
\value{
A vector of seleced stocks
}
\description{
To search for stocks that was sold off and then becomes stable.
Watch them for reversal.
}
