% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_own_ETF.R
\name{plot_etf}
\alias{plot_etf}
\title{Plot the chart of self-defined ETF}
\usage{
plot_etf(etf_stocks, etf_name = "ETF")
}
\arguments{
\item{etf_stocks}{vector of stock symbols that are components of the ETF}

\item{eft_name}{name of the ETF}
}
\description{
Plot the chart of self-defined ETF
}
