% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_revenue_growth}
\alias{plot_revenue_growth}
\title{Plot revenue growth of stocks}
\usage{
plot_revenue_growth(stocks)
}
\arguments{
\item{stocks}{vector of stock symbols such as c("C", "AAPL")}
}
\description{
Plot revenue growth of stocks
}
