% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trader_work_station_functions.R
\name{plot_tws}
\alias{plot_tws}
\title{chart stocks using tws data}
\usage{
plot_tws(stocks, period = 300, description = NULL)
}
\arguments{
\item{stocks}{vector of stock symbols such as c("C", "AAPL")}

\item{period}{number of trading day to display}

\item{description}{description of the chart, stock symbol by default}
}
\description{
chart stocks using tws data
}
