% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stock_selection.R
\name{select_price}
\alias{select_price}
\title{Select from a pool of stocks whose last trading price are in the given range}
\usage{
select_price(stocks, from, to)
}
\arguments{
\item{stocks}{a subset of all stocks generated by \link{get_all_stocks}}

\item{from, to}{lower and upper limits of last trading price to be selected.}
}
\value{
a data.table
}
\description{
Select from a pool of stocks whose last trading price are in the given range
}
