% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_exchange}
\alias{stock_exchange}
\title{US stocks and exchanges}
\format{A data.table with 6896 rows and 2 variables
\describe{
  \item{symbol}{stock symbol}
  \item{exchange}{exchange for the stock}
}}
\usage{
stock_exchange
}
\description{
US stocks and exchanges
}
\keyword{datasets}
