% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stock_selection.R
\name{tq2qm}
\alias{tq2qm}
\title{convert tidyquant::tq_get data() to quantmod::getSymbols() data}
\usage{
tq2qm(tq_df)
}
\arguments{
\item{dq_df}{a data frame downloaded by tidyquant::tq_get()}
}
\value{
a data.frame of format getting by quantmod::getSymbols()
}
\description{
convert tidyquant::tq_get data() to quantmod::getSymbols() data
}
